display result on FE
