+++ app.js
@@ -17,24 +17,24 @@ const MongoStore = require("connect-mongo")(session);
 const ensureLogin = require('connect-ensure-login');
 require("./config/passport")
 mongoose
- .connect('mongodb://localhost/bres', { useNewUrlParser: true })
- .then(x => {
-   console.log('Connected to Mongo! Database name: "${x.connections[0].name}"')
- })
- .catch(err => {
-   console.error('Error connecting to mongo', err)
- });
+  .connect('mongodb://localhost/bres', { useNewUrlParser: true })
+  .then(x => {
+    console.log('Connected to Mongo! Database name: "${x.connections[0].name}"')
+  })
+  .catch(err => {
+    console.error('Error connecting to mongo', err)
+  });
 
 
 const app_name = require('./package.json').name;
 const debug = require('debug')(`${app_name}:${path.basename(__filename).split('.')[0]}"`);
 const app = express();
 app.use(session({
- secret: "abc",
- store: new MongoStore({
-   mongooseConnection: mongoose.connection,
-   ttl: 24 * 60 * 60 // 1 day
- })
+  secret: "abc",
+  store: new MongoStore({
+    mongooseConnection: mongoose.connection,
+    ttl: 24 * 60 * 60 // 1 day
+  })
 }));
 // Middleware Setup
 app.use(logger('dev'));
@@ -46,9 +46,9 @@ app.use(passport.initialize())
 app.use(passport.session())
 // Express View engine setup
 app.use(require('node-sass-middleware')({
- src: path.join(__dirname, 'public'),
- dest: path.join(__dirname, 'public'),
- sourceMap: true
+  src: path.join(__dirname, 'public'),
+  dest: path.join(__dirname, 'public'),
+  sourceMap: true
 }));
 app.set('views', path.join(__dirname, 'views'));
 app.set('view engine', 'hbs');
@@ -63,9 +63,13 @@ app.use('/', index);
 const indexRouter = require('./routes/index')
 const usersRouter = require('./routes/users')
 const authRouter = require('./routes/auth')
-const noteRouter = require('./routes/home')
+const homeRouter = require('./routes/home')
+const resultsRouter = require('./routes/results')
+
 app.use('/', indexRouter)
 app.use('/', authRouter)
 app.use('/users', usersRouter)
-app.use('/home', ensureLogin.ensureLoggedIn(), noteRouter)
+app.use('/home', ensureLogin.ensureLoggedIn(), homeRouter)
+app.use('/', ensureLogin.ensureLoggedIn(), resultsRouter)
+
 module.exports = app;
