+++ public/stylesheets/style.css
@@ -6,74 +6,75 @@ body {
   font: 14px "Lucida Grande", Helvetica, Arial, sans-serif;
   flex-wrap: wrap;
 }
+#service {
+  margin-top: 200px;
+}
+.resultImage {
+  width: 100px;
+  border-radius: 10%;
+}
 
+.container {
+  display: flex;
+  width: 100%;
+  background-color: #2c2c2c;
+  position: fixed;
+  top: 0;
+}
 
-  .container{
-    display: flex;
-    width: 100%;
-    background-color: #2c2c2c;
-    position: fixed;
-    top: 0;
-    
-  }
+.menu1 {
+  width: 100%;
+  padding: auto;
+  margin: auto;
+}
 
-  .menu1{
-    width: 100%;
-    padding: auto;
-    margin: auto;
-  }
+.menu1 ul {
+  display: flex;
+  justify-content: flex-start;
+  padding: auto;
+  margin-left: 0px;
+}
 
-    .menu1 ul{
-      display: flex;
-      justify-content: flex-start;
-      padding: auto;
-      margin-left: 0px;
-      
-    }
-   
-  .menu2{
-    width: 100%;
-    padding: auto;
-    margin: auto;
-    
-  }
-  .menu2 ul{
-    display: flex;
-    justify-content: flex-end;
-    padding: auto;
-    margin-right: 50px;
-    
-  }
+.menu2 {
+  width: 100%;
+  padding: auto;
+  margin: auto;
+}
+.menu2 ul {
+  display: flex;
+  justify-content: flex-end;
+  padding: auto;
+  margin-right: 50px;
+}
 
-  .navigation a {
-    text-decoration: none;
-    font: 16px "Lucida Grande", Helvetica, Arial, sans-serif;
-    display: block;
-    padding: auto;
-    color: #ffd000;;;
-  }
+.navigation a {
+  text-decoration: none;
+  font: 16px "Lucida Grande", Helvetica, Arial, sans-serif;
+  display: block;
+  padding: auto;
+  color: #ffd000;
+}
 
-  .navigation a:hover {
-    color: #ffffff;
-  }
+.navigation a:hover {
+  color: #ffffff;
+}
 
-  li {
+li {
   display: inline;
   padding: 0px 10px;
   list-style-type: none;
-  }
-
-  .banner img{
-    display: flex;
-    height: 100%;
-    width: 100%;
 }
 
-  .search {
-    position: absolute;
-    margin-top: -210px;
-    margin-left: 300px;
-    
+.banner img {
+  display: flex;
+  height: 100%;
+  width: 100%;
+}
+
+.search {
+  position: absolute;
+  margin-top: -210px;
+  margin-left: 300px;
 }
 
 input::placeholder {
@@ -81,14 +82,13 @@ input::placeholder {
   font-size: 25px;
 }
 
-  .searchBar {
-    height: 35px;
-    width: 350px;
-    border: 1px solid #ffffff;
-    border-radius: 10px;
-    outline: none;
-    background-color: white;
-
+.searchBar {
+  height: 35px;
+  width: 350px;
+  border: 1px solid #ffffff;
+  border-radius: 10px;
+  outline: none;
+  background-color: white;
 }
 
 .location {
@@ -104,29 +104,27 @@ input::placeholder {
   border-radius: 10px;
   outline: none;
   background-color: white;
-
 }
 
- .login{
-    width: auto;
-    padding: 10px;
-    margin-top: 100px;
-    margin-left: 400px;
-    background-color: #ffffff;
-    position: fixed;
+.login {
+  width: auto;
+  padding: 10px;
+  margin-top: 100px;
+  margin-left: 400px;
+  background-color: #ffffff;
+  position: fixed;
 }
 
-.login input{
-    height: 35px;
-    width: 350px;
-    border: none;
-    border-bottom: 2px solid #2c2c2c;
-    outline: none;
-    border-radius: 10px;
-  
+.login input {
+  height: 35px;
+  width: 350px;
+  border: none;
+  border-bottom: 2px solid #2c2c2c;
+  outline: none;
+  border-radius: 10px;
 }
 
-button{
+button {
   height: 40px;
   width: auto;
   padding: 10px;
@@ -139,7 +137,7 @@ button{
   border: none;
 }
 
-.goButton{
+.goButton {
   height: 40px;
   width: auto;
   padding: 10px;
@@ -151,47 +149,38 @@ button{
   border-radius: 10px;
 }
 
-  
-  .footer{
-    display: flex;
-    width: 100%;
-    background-color: #2c2c2c;
-    position: fixed;
-    bottom: 0%;
-   
-    
-  }
+.footer {
+  display: flex;
+  width: 100%;
+  background-color: #2c2c2c;
+  position: fixed;
+  bottom: 0%;
+}
 
-  .menu3{
-    width: auto;
-    padding: auto;
-    margin-left: 10px;
-  }
+.menu3 {
+  width: auto;
+  padding: auto;
+  margin-left: 10px;
+}
 
-    .menu3 ul{
-      display: flex;
-      flex-direction: column;
-      padding: auto;
-      margin-left: 0px;
-      
-    }
+.menu3 ul {
+  display: flex;
+  flex-direction: column;
+  padding: auto;
+  margin-left: 0px;
+}
 
-    .flex-container {
-     display: flex;
-     flex-direction: row;
-     
-      
-    }
+.flex-container {
+  display: flex;
+  flex-direction: row;
+}
 
-    .item {
-      margin: 10px;
-      display: inherit;
-      padding: 10px;
-      width: 100px;
-      height: 100px;
-      align-items: center;
-      background-color:#ffd000;
-      
-    }
-
-    
+.item {
+  margin: 10px;
+  display: inherit;
+  padding: 10px;
+  width: 100px;
+  height: 100px;
+  align-items: center;
+  background-color: #ffd000;
+}
