+++ routes/home.js
@@ -1,10 +1,11 @@
 const express = require('express');
-const router  = express.Router();
+const router = express.Router();
 
 // GET Home page
 
 router.get('/', (req, res, next) => {
   res.render('home');
+
 });
 
 module.exports = router;
