+++ views/home.hbs
@@ -1,40 +1,41 @@
 <div class="main">
-  
   <div class="banner">
-  <img src="images/banner1.jpg">
+    <img src="images/banner1.jpg">
+  </div>
+  <form action="/results" method="GET">
+
+    <div class="search">
+      <input class="searchBar" name="category" type="text" autocomplete="on" placeholder="Search...">
+    </div>
+
+    <div class="location">
+      <input name="location" class="locationBar" type="text" autocomplete="on" placeholder="Where?">
+      <button class="goButton" type="submit">Go!</button>
+    </div>
+  </form>
+
+</div>
+
+
+
+<div class="flex-container" id="justify">
+  <div class="item" data-color="2a80b9">
+    <p>1</p>
   </div>
 
-  <div class="search">
-  <input class="searchBar" type="text" autocomplete="on" placeholder="Search...">
+  <div class="item" data-color="2a80b9">
+    <p>1</p>
+  </div>
+
+  <div class="item" data-color="2a80b9">
+    <p>1</p>
+  </div>
+
+  <div class="item" data-color="2a80b9">
+    <p>1</p>
+  </div>
+
+  <div class="item" data-color="2a80b9">
+    <p>1</p>
+  </div>
 </div>
-
- <div class="location">
- <input class="locationBar" type="text" autocomplete="on" placeholder="Where?">
- <button class="goButton" type="submit">Go!</button>
-</div>
-
-</div>
-
-
-
- <div class="flex-container" id="justify">
-    <div class="item" data-color="2a80b9">
-      <p>1</p>
-    </div>
-
-    <div class="item" data-color="2a80b9">
-      <p>1</p>
-    </div>
-
-    <div class="item" data-color="2a80b9">
-      <p>1</p>
-    </div>
-
-    <div class="item" data-color="2a80b9">
-      <p>1</p>
-    </div>
-
-    <div class="item" data-color="2a80b9">
-      <p>1</p>
-    </div>
-    </div>
